+++ src/pages/Login/LoginPage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LoginForm from "../../components/LoginForm";
+import LoginForm from "../../components/LoginForm/LoginForm";
 import "./LoginPage.scss";
 
 const LoginPage: React.FC = () => {
